Improve `insert_pipeline_barriers` performance
This function takes a significant amount of time each frame. Since it is
called often adding a bump allocator for scratch memory reduces time
spent allocating/deallocating significantly.

We'll also make the function `unsafe` to remove some validity checks
which take a combined ~10-15% of time spent in that function. The
validity requirements are already correctly handlded by the executor.